Map: optimize PathFinder circle calculations
use our PlotCircle function instead of brute force with wasted iterations and 4 extra Point calculations (this might help out with #1692)

The originals probably used an approximation (probably also Bresenham's) so maybe we are now truer to the originals, changing comment about the difference into a TODO to recheck

we can also eliminate a redundant bounds check because QuerySearchMap already does this

+++ gemrb/core/Map.h
@@ -373,7 +373,7 @@ public:
 	Color QueryLighting(const Point& p) const noexcept;
 	
 	void PaintSearchMap(const Point&, PathMapFlags value) const noexcept;
-	void PaintSearchMap(const Point& Pos, unsigned int blocksize, PathMapFlags value) const noexcept;
+	void PaintSearchMap(const Point& Pos, uint16_t blocksize, PathMapFlags value) const noexcept;
 };
 
 class GEM_EXPORT Map : public Scriptable {
@@ -725,7 +725,7 @@ private:
 	// same as GetBlocked, but in TileCoords
 	PathMapFlags GetBlockedTile(const Point&) const;
 	PathMapFlags GetBlockedTile(const Point&, int size) const;
-	PathMapFlags GetBlockedInRadiusTile(const Point&, unsigned int size, bool stopOnImpassable = true) const;
+	PathMapFlags GetBlockedInRadiusTile(const Point&, uint16_t size, bool stopOnImpassable = true) const;
 };
 
 }
